+++ src/js/validator/hexColor.js
@@ -18,7 +18,11 @@
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
-            return $.fn.bootstrapValidator.validators.color._hexColor(validator, $field, options);
+            var value = $field.val();
+            if (value === '') {
+                return true;
+            }
+            return /(^#[0-9A-F]{6}$)|(^#[0-9A-F]{3}$)/i.test(value);
         }
     };
 }(window.jQuery));